Ready to be tested i think

+++ aimbots-src/src/informants/communication/communication_message.hpp
@@ -5,19 +5,19 @@ namespace src::Communication {
 #ifdef TARGET_SENTRY
 struct robot_state_message_team {
     /* data */
-    uint32_t standardX, standardY, heroX, heroY, sentryX, sentryY;
+    uint16_t standardX, standardY, heroX, heroY, sentryX, sentryY;
 };
 
 #else
 struct robot_state_message {
     /* data */
     uint8_t robotID;
-    uint32_t x, y;
+    uint16_t x, y;
 };
 
 struct robot_state_message_enemy {
     uint8_t robotID;
-    uint32_t x, y;
+    uint16_t x, y;
     uint16_t timestamp;
 };
